[Generator] Make generator.cs smaller by not using nested enums.
Code is complicated, lets remove as much noise as possible to focus on
the important parts.
