+++ src/bgen/ThreadCheck.cs
@@ -1,0 +1,5 @@
+public enum ThreadCheck {
+	Default, // depends on the namespace
+	Off,
+	On,
+}
